+++ src/index.js
@@ -6,6 +6,7 @@ import Icon from './Icon';
 import StatusIcon from './StatusIcon';
 import Tag from './Tag';
 import Text from './Text';
+import Tooltip from './Tooltip';
 
 // Layout helpers
 import FlexCell from './FlexCell';
@@ -24,6 +25,7 @@ export {
     StatusIcon,
     Tag,
     Text,
+    Tooltip,
 
     FlexCell,
     IconLayout,