+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   * Generate Cobertura coverage report.
   * Run `npm publish` with `NPM_TOKEN` env variables in `deploy.sh`.
 - `<Tag>` now adapts to the color of its parent more actively.
+- Fix `<Button>` should turn red on error state
+- `wrapIfNotElement()` helper now takes an extra `via` param to change how `content` is passed to the `Wrapper` component.
 
 ## [0.5.0]
 ### Added