+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+
+## [0.6.0]
+### Added
 - Migrate row components:
   * `<Button>`
 - Migrate visual elements: